Separated kernel status and key value into components
