+++ src/components/KeyValue.tsx
@@ -1,0 +1,45 @@
+import React from 'react';
+import { StyleSheet, css } from 'aphrodite';
+
+import { spacing } from '../constants/theme';
+
+const styles = StyleSheet.create({
+  keyValue: {
+    width: '100%',
+    display: 'flex',
+    flexDirection: 'column',
+    marginBottom: spacing.DEFAULT / 2,
+  },
+  keyText: {
+    width: '100%',
+    display: 'flex',
+    flexDirection: 'row',
+    alignItems: 'center',
+    fontWeight: 'bold',
+    textTransform: 'uppercase',
+    fontSize: 10,
+  },
+  valueText: {
+    width: '100%',
+    display: 'flex',
+    flexDirection: 'row',
+    alignItems: 'center',
+  },
+});
+
+/**
+ * A key value pair component
+ */
+const KeyValue: React.FC<{ attributeKey: string | React.ReactNode; attributeValue: string | React.ReactNode }> = ({
+  attributeKey,
+  attributeValue,
+}) => {
+  return (
+    <div className={css(styles.keyValue)}>
+      <span className={css(styles.keyText)}>{attributeKey}</span>
+      <span className={css(styles.valueText)}>{attributeValue}</span>
+    </div>
+  );
+};
+
+export default KeyValue;
